% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_entries_to_matrix.R
\name{add_entries_to_matrix}
\alias{add_entries_to_matrix}
\title{Add entries to the matrix based on the genome pairs}
\usage{
add_entries_to_matrix(df)
}
\arguments{
\item{df}{Input data frame containing genomes as columns.}
}
\value{
Data frame with pairwise gene sums.
}
\description{
Add entries to the matrix based on the genome pairs
}
\examples{
# Example data
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")

initial_matrix <- create_initial_matrix(cd_hit_file)
new_df <- add_entries_to_matrix(initial_matrix)
}
\seealso{
\code{\link[=create_initial_matrix]{create_initial_matrix()}}
}
