% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_matrix.R
\name{create_initial_matrix}
\alias{create_initial_matrix}
\title{Create an initial matrix from a genome list file}
\usage{
create_initial_matrix(file_path)
}
\arguments{
\item{file_path}{Path to the genome list file.}
}
\value{
Returns initial matrix with rows and columns corresponding to unique genome IDs.
}
\description{
Create an initial matrix from a genome list file
}
\examples{
# Example data
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")

initial_matrix <- create_initial_matrix(cd_hit_file)
}
