% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorGeneVis-package.R
\docType{package}
\name{HorGeneVis-package}
\alias{HorGeneVis}
\alias{HorGeneVis-package}
\title{HorGeneVis: Horizontal Gene Transfer Visualization}
\description{
This package involves the analysis and visualization of horizontal gene transfer (HGT) between individual taxonomic categories using genomic data. The tool combines the integration of a heat-map matrix with phylogenetic analysis visualization techniques to enhance the accuracy and efficiency of HGT event detection. We applied it to a diverse set of bacterial genomes from chicken and pig gut and expanded the results with established methods for HGT detection. The findings demonstrate the capability of our visualization techniques to pinpoint HGT events, offering a valuable tool for researchers to explore the intricate dynamics of bacterial genome evolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JanaSchwarzerova/HorGeneVis}
  \item Report bugs at \url{https://github.com/JanaSchwarzerova/HorGeneVis/issues}
}

}
\author{
\strong{Maintainer}: Jana Schwarzerova \email{Jana.Schwarzerova@vut.cz} (\href{https://orcid.org/0000-0003-2918-9313}{ORCID})

Authors:
\itemize{
  \item Laxmipriya Rajasekaran (\href{https://orcid.org/0009-0000-6401-7731}{ORCID})
  \item Katerina Jureckova (\href{https://orcid.org/0000-0001-5143-9521}{ORCID})
  \item Julie Nejezchlebova (\href{https://orcid.org/0009-0008-2290-6467}{ORCID})
  \item Margaret Varga (\href{https://orcid.org/0000-0002-9086-1626}{ORCID})
  \item Valentine Provaznik (\href{https://orcid.org/0000-0002-3422-7938}{ORCID})
  \item Wolfram Weckwerth (\href{https://orcid.org/0000-0002-9719-6358}{ORCID})
  \item Darina Cejkova (\href{https://orcid.org/0000-0002-6989-6330}{ORCID}) [funder]
}

}
\keyword{internal}
