% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{HorGeneVis_example_data}
\alias{HorGeneVis_example_data}
\title{Get path to HorGeneVis example data}
\usage{
HorGeneVis_example_data(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
HorGeneVis comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
HorGeneVis_example_data()
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")
tree_file <- HorGeneVis_example_data("PhylogeneticTree_newick.txt")
tree_list_file <- HorGeneVis_example_data("Tree_list_Rectangular_output_IToL.txt")
}
