% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_final_matrix.R
\name{process_final_matrix}
\alias{process_final_matrix}
\title{Process the final matrix based on a tree list file}
\usage{
process_final_matrix(tree_file, mFin_ff)
}
\arguments{
\item{tree_file}{Path to the tree list file.}

\item{mFin_ff}{Input matrix to process.}
}
\value{
Processed final matrix.
}
\description{
Process the final matrix based on a tree list file
}
\examples{
# Example data
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")
tree_list_file <- HorGeneVis_example_data("Tree_list_Rectangular_output_IToL.txt")

initial_matrix <- create_initial_matrix(cd_hit_file)
new_df <- add_entries_to_matrix(initial_matrix)
final_mFin_ff_p <- process_final_matrix(tree_list_file, new_df)
sum(final_mFin_ff_p)
}
\seealso{
\code{\link[=create_initial_matrix]{create_initial_matrix()}} \code{\link[=add_entries_to_matrix]{add_entries_to_matrix()}}
}
