% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tree.R
\name{process_tree}
\alias{process_tree}
\title{Process a phylogenetic tree file and extract new names}
\usage{
process_tree(tree_file, mFin_T)
}
\arguments{
\item{tree_file}{Path to the phylogenetic tree file.}

\item{mFin_T}{Matrix where the column names are to be matched. The output from create_initial_matrix() function.}
}
\value{
Vector of new names extracted from the tree file.
}
\description{
Process a phylogenetic tree file and extract new names
}
\examples{
Example data
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")
tree_file <- HorGeneVis_example_data("PhylogeneticTree_newick.txt")

initial_matrix <- create_initial_matrix(cd_hit_file)
new_names <- process_tree(tree_file, initial_matrix)
}
\seealso{
\code{\link[=create_initial_matrix]{create_initial_matrix()}}
}
