% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chord_diagram.R
\name{create_chord_diagram}
\alias{create_chord_diagram}
\title{Create a chord diagram with default settings}
\usage{
create_chord_diagram(subset_new_df, transparency = 0.5)
}
\arguments{
\item{subset_new_df}{Subset of the input data frame for creating the chord diagram.}

\item{transparency}{Transparency level for the chord diagram (default is 0.5).}
}
\value{
A data frame which contains positions of links in Chord Diagram
}
\description{
Create a chord diagram with default settings
}
\examples{
# Example data
cd_hit_file <- HorGeneVis_example_data("HGT_output_CD_hit.txt")

initial_matrix <- create_initial_matrix(cd_hit_file)
new_df <- add_entries_to_matrix(initial_matrix)
create_chord_diagram(new_df[1:15, 16:20])
}
\seealso{
\code{\link[=create_initial_matrix]{create_initial_matrix()}} \code{\link[=add_entries_to_matrix]{add_entries_to_matrix()}}
}
