% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorGeneVis-package.R
\docType{package}
\name{HorGeneVis-package}
\alias{HorGeneVis}
\alias{HorGeneVis-package}
\title{HorGeneVis: Horizontal Gene Transfer Visualization}
\description{
This package involves the analysis and visualization of horizontal gene transfer (HGT) between individual taxonomic categories using genomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JanaSchwarzerova/HorGeneVis}
  \item Report bugs at \url{https://github.com/JanaSchwarzerova/HorGeneVis/issues}
}

}
\author{
\strong{Maintainer}: Jana Schwarzerova \email{Jana.Schwarzerova@vut.cz}

Authors:
\itemize{
  \item Laxmipriya Rajasekaran
  \item Katerina Jureckova
  \item Julie Nejezchlebova
  \item Margaret Varga
  \item Valentine Provaznik
  \item Wolfram Weckwerth
  \item Darina Cejkova [funder]
}

}
\keyword{internal}
